+++ api/serializers.py
@@ -1,0 +1,37 @@
+﻿from rest_framework import serializers
+from businesses.models import Business
+from services.models import Service
+from appointments.models import Appointment
+from locations.models import BusinessLocation
+from staff.models import StaffMember
+from django.contrib.auth.models import User
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ['id', 'username', 'first_name', 'last_name', 'email']
+
+class BusinessSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Business
+        fields = '__all__'
+
+class ServiceSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Service
+        fields = '__all__'
+
+class StaffMemberSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = StaffMember
+        fields = '__all__'
+
+class AppointmentSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Appointment
+        fields = '__all__'
+
+class BusinessLocationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = BusinessLocation
+        fields = '__all__'